refactor name change to 'Dog'
