+++ dog/skill.properties
@@ -1,5 +1,6 @@
-name = FurhatDog
-mainclass = furhatos.app.furhatdog.FurhatdogSkill
+name = Dog
+mainclass = furhatos.app.dog.DogSkill
+version = 1.0.0
 language = en-US
 logLevel = INFO
 #You may set this to a Furhat version.
