+++ dog/src/main/kotlin/furhatos/app/dog/extensions/flowControlRunnerExtensions.kt
@@ -1,15 +1,28 @@
-package furhatos.app.furhatdog.extensions
+package furhatos.app.dog.extensions
 
-import furhatos.app.furhatdog.dog.CurrentState
-import furhatos.app.furhatdog.dog.Dog
-import furhatos.app.furhatdog.gestures.*
+import furhatos.app.dog.dog.CurrentState
+import furhatos.app.dog.dog.Dog
+import furhatos.app.dog.flow.MissingMask
+import furhatos.app.dog.gestures.*
 import furhatos.flow.kotlin.FlowControlRunner
 import furhatos.flow.kotlin.RandomPolicy
 import furhatos.flow.kotlin.furhat
 import gestures.WakeUpWithHeadShake
 
 fun FlowControlRunner.setDogCharacter() {
-    furhat.setModel("pug", "default")
+    if (furhat.faces.get("pug") != null) {
+        furhat.setModel("pug", "default")
+    } else if (furhat.faces.get("dog") != null) {
+        furhat.setModel("dog", "default")
+    } else {
+        furhat.say {
+            +"This skill requires that you have the dog mask and the dog 3D model. "
+            +"So you have to fix that, before I can take the shape of the dog called Lucky. "
+            +delay(400)
+            +"Go to the home tab, in the web console, to see what mask models are currently available for me. "
+        }
+        goto(MissingMask)
+    }
 }
 
 fun FlowControlRunner.randomBark() {
@@ -50,7 +63,7 @@ fun FlowControlRunner.randomGrowl() {
 }
 
 fun FlowControlRunner.awakeIfAsleep(userId: String) {
-    when(Dog.currentState) {
+    when (Dog.currentState) {
         CurrentState.AWAKE -> {
             furhat.attend(userId)
         }
