+++ dog/src/main/kotlin/furhatos/app/dog/flow/responseHandlers.kt
@@ -1,28 +1,21 @@
-package furhatos.app.furhatdog.flow
+package furhatos.app.dog.flow
 
-import furhatos.app.furhatdog.dog.CurrentState
-import furhatos.app.furhatdog.dog.Dog
-import furhatos.app.furhatdog.extensions.awakeIfAsleep
-import furhatos.app.furhatdog.extensions.randomBark
-import furhatos.app.furhatdog.extensions.randomGrowl
-import furhatos.app.furhatdog.extensions.randomWhimpering
-import furhatos.app.furhatdog.gestures.bark1
-import furhatos.app.furhatdog.gestures.growl2
-import furhatos.app.furhatdog.gestures.growl4
-import furhatos.app.furhatdog.gestures.panting1
-import furhatos.app.furhatdog.gestures.shake1
-import furhatos.app.furhatdog.gestures.whimpering1
-import furhatos.app.furhatdog.gestures.whimpering2
-import furhatos.app.furhatdog.gestures.whimpering3
-import furhatos.app.furhatdog.gestures.whimpering5
-import furhatos.app.furhatdog.nlu.BadDog
-import furhatos.app.furhatdog.nlu.Fetch
-import furhatos.app.furhatdog.nlu.GoodDog
-import furhatos.app.furhatdog.nlu.ILoveYou
-import furhatos.app.furhatdog.nlu.LayDown
-import furhatos.app.furhatdog.nlu.RollAround
-import furhatos.app.furhatdog.nlu.SitDown
-import furhatos.app.furhatdog.nlu.YouAreCute
+import furhatos.app.dog.dog.CurrentState
+import furhatos.app.dog.dog.Dog
+import furhatos.app.dog.extensions.awakeIfAsleep
+import furhatos.app.dog.extensions.randomBark
+import furhatos.app.dog.extensions.randomGrowl
+import furhatos.app.dog.extensions.randomWhimpering
+import furhatos.app.dog.gestures.panting1
+import furhatos.app.dog.gestures.shake1
+import furhatos.app.dog.nlu.BadDog
+import furhatos.app.dog.nlu.Fetch
+import furhatos.app.dog.nlu.GoodDog
+import furhatos.app.dog.nlu.ILoveYou
+import furhatos.app.dog.nlu.LayDown
+import furhatos.app.dog.nlu.RollAround
+import furhatos.app.dog.nlu.SitDown
+import furhatos.app.dog.nlu.YouAreCute
 import furhatos.flow.kotlin.RandomPolicy
 import furhatos.flow.kotlin.furhat
 import furhatos.flow.kotlin.onResponse
