+++ dog/src/main/kotlin/furhatos/app/dog/gestures/lookBackAndAway.kt
