--- dog/src/main/kotlin/furhatos/app/furhatdog/users.kt
@@ -1,2 +1,0 @@
-package furhatos.app.furhatdog.flow
-
